Add handling of flipable pieces
I noticed that there was one piece that could be laid down with either
side facing up, so I thought that I would have to add handling of that
to be able to solve all possible combinations.

I added a test that will run all combinations of months, days and
weekdays to be sure my code works. I have a rough estimate that the test
will take up to 90 minutes to complete, so I have not ran it all the way
through yet.